Platform/RaspberryPi4: Add a basic thermal zone
Rather than exporting the temp sensor or mailbox
in ACPI land we can wrap them in AML and use the default
ACPI drivers provided by the OS.

This commit adds a basic active cooling ACPI thermalzone
with trip points for active cooling (fan), passive cooling
 (throttling), hibernate and critical shutdown. The
hibernate+critical trip points probably need further
tweaking (hibernate should probably be ~85, with crit
a couple degree's higher), as the vc firmware is doing
the throtting for us at the moment (that should somehow be
reflected via CPPC/etc in the future). Apparently this
throtting kicks in ~80C.

The fan control is currently hardcoded to GPIO pin 19, and
is a basic on/off control rather than using the pwm function
the pin provides. The assumption is that the user has
wired up GPIO fan control as is documented in various places
on the internet using a transistor/whatever to provide a higher
current source for the fan, leaving the 3.3V GPIO as only
control. It is possible with smaller fans to do this all with
the5V supply + GPIO pin.

Signed-off-by: Jeremy Linton <lintonrjeremy@gmail.com>