rpi4/acpi: Add PCIe SSDT
Since we plan on toggling between XHCI and PCI the PCI
root needs to be in its own SSDT. This is all thats needed
of UEFI. The SMC conduit is provided directly to the running
OS. When the OS detects this PCIe port, on a machine without
a MADT it attempts to connect to the SMC conduit. This definition
doesn't have any power mgmt, and only provides a description of
the root port.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>