Platform/RaspberryPi/Acpitables: Add emmc2 device and modify existing CID
The current arasan controller actually mostly works in linux as a
PNP0D40 (SDHCI) compatible device. All we seemingly need to do is
add an override for the capabilities registers which are unhelpfully
zero'ed on the rpi4, rather than representing the device capabilities.
A large part of that is likely because its running in "PIO" mode
rather than using any of the more advanced DMA modes, which also
means its quite slow.

The newer emmc2 controller also seems to mostly run as a PNP0D40
if we disable ADMA2 in the caps register. Although it currently
has a problem that the CMD12 quirk isn't set so linux spews
a lot of unexpected interrupt messages, but supprisingly
continues to work sufficiently that if the BDS/config
is changed to emmc2, both the SD card and wifi are somewhat
functional (the latter needs the firmware .txt file linked
to the rpi4 version, as its using our SMBIOS name rather
than the dt machine name, that should be something we can
upstream).

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>